Add six-pack feature
Bam. Pattern matching once again. I'm not sure if this is the ideal
solution. Compared to the OOP version, it's not clear that there's a
BottleNumber concept. But I'm not sure if unearthing that concept
matters because using pattern matching, as soon as we DRY'ed the verses,
the code was open to six packs.